pyproject.toml
[project]
name = "seatalk-search"
version = "0.1.0"
description = "MCP Server for semantic search of SeaTalk conversations"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Aloy", email = "aloysius.chia@shopee.com"},
]

dependencies = [
    "fastmcp>=2.6.0",
    "sentence-transformers>=3.2.0",
    "numpy>=1.24.0,<2.0.0",
    "torch>=2.0.0",
    "apsw-sqlite3mc>=0.9.10",
    "tqdm>=4.66.0",
    "python-dateutil>=2.8.0",
    "cryptography>=3.4.8",
]

[project.scripts]
seatalk-search-server = "src.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"] 
